Move die() and warn() from bin/__init__.py to core.py.
